% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_plots_faib.R
\name{faib_leading_species_code}
\alias{faib_leading_species_code}
\title{Extract the leading species' code from a composition string}
\usage{
faib_leading_species_code(x)
}
\arguments{
\item{x}{Character vector of \code{SPB_CPCT_LS} composition strings.}
}
\value{
Character vector of two-letter species codes.
}
\description{
Extract the leading species' code from a composition string
}
