% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuel_typing.R
\name{normalize_fbp_codes}
\alias{normalize_fbp_codes}
\title{Collapse bcwsft season/leaf fuel-type codes to the provincial Public vocabulary}
\usage{
normalize_fbp_codes(x)
}
\arguments{
\item{x}{Character vector of bcwsft fuel-type codes.}
}
\value{
A character vector with the seasonal/structural leaves collapsed to
the aggregated provincial codes.
}
\description{
bcwsft emits seasonal/structural leaves (D-1/D-2, M-1/M-2, O-1a/O-1b) that the
provincial layer reports aggregated (D-1/2, M-1/2, O-1a/b).
}
\seealso{
Other BC fire and fuel data:
\code{\link[=calc_recently_disturbed]{calc_recently_disturbed()}},
\code{\link[=clip_nfdb_to_study_area]{clip_nfdb_to_study_area()}},
\code{\link[=compare_fuel_typing]{compare_fuel_typing()}},
\code{\link[=fuel_types_distribution]{fuel_types_distribution()}},
\code{\link[=get_fuel_types]{get_fuel_types()}},
\code{\link[=get_vri_for_fuel_typing]{get_vri_for_fuel_typing()}},
\code{\link[=load_nbac_polys]{load_nbac_polys()}},
\code{\link[=load_nfdb_points]{load_nfdb_points()}},
\code{\link[=load_nfdb_polys]{load_nfdb_polys()}},
\code{\link[=plot_fuel_typing_comparison]{plot_fuel_typing_comparison()}},
\code{\link[=prep_fuel_types_rast]{prep_fuel_types_rast()}},
\code{\link[=run_bcwsft_fuel_typing]{run_bcwsft_fuel_typing()}}
}
\concept{BC fire and fuel data}
